refactor(config): centralize site config in config.json and update references
- Added src/_data/config.json and config.mjs for centralized configuration
- Removed redundant config from site.json and site.mjs
- Updated templates and layouts to use config instead of site for config values
- Adjusted image widths, keys, and URLs to reference config
- Improved maintainability by consolidating config data and usage
