+++ src/_includes/category-list-item.html
@@ -1,7 +1,7 @@
 <li>
   {% assign item = category %}
-  {% assign widths = site.products.category_thumb_widths %}
-  {% assign sizes = site.products.category_thumb_sizes %}
+  {% assign widths = config.products.category_thumb_widths %}
+  {% assign sizes = config.products.category_thumb_sizes %}
   {% include "thumbnail-image-link.html" %}
 
   {% if category.data.description -%}
