+++ src/_includes/gallery.html
@@ -1,12 +1,12 @@
 <div class="current-image">
   {% assign first_image = gallery[0] %}
-  {% image first_image.filename, first_image.alt, site.products.gallery_image_widths, "", "100vw" %}
+  {% image first_image.filename, first_image.alt, config.products.gallery_image_widths, "", "100vw" %}
 </div>
 
 <div aria-hidden="true" class="gallery-full-size-images">
   {%- for image in gallery -%}
     <div class="full-image-{{ forloop.index }}">
-      {% image image.filename, image.alt, site.products.gallery_image_widths, "fade-in", "100vw" %}
+      {% image image.filename, image.alt, config.products.gallery_image_widths, "fade-in", "100vw" %}
     </div>
   {% endfor %}
 </div>
@@ -27,7 +27,7 @@
       >
         <!-- once sizes="auto" is supported by Firefox, switch to that -->
         <figure>
-          {% image image.filename, "", site.products.gallery_thumb_widths, "", "(max-width: 798px) 33vw, 25vw" %}
+          {% image image.filename, "", config.products.gallery_thumb_widths, "", "(max-width: 798px) 33vw, 25vw" %}
           {%- if image.alt != "" -%}
             <figcaption>{{- image.alt -}}</figcaption>
           {%- endif -%}
