+++ src/_includes/head.html
@@ -28,14 +28,14 @@
     src="/assets/js/lib/turbo.es2017-esm.min.js"
   ></script>
 
-  {%- if stripe_buy_button_id and site.stripe_publishable_key -%}
+  {%- if stripe_buy_button_id and config.stripe_publishable_key -%}
     <script
       async
       src="https://js.stripe.com/v3/buy-button.js"
     ></script>
   {%- endif -%}
 
-  {%- if layout == "contact.html" and site.botpoison_public_key and site.botpoison_public_key != "" -%}
+  {%- if layout == "contact.html" and config.botpoison_public_key and config.botpoison_public_key != "" -%}
     <script
       async
       src="/assets/js/lib/botpoison-0.1.30.js"
