+++ src/_includes/header.html
@@ -7,10 +7,10 @@
       {{ header_image | default: page.header_image }}
     {%- endcapture %}
     <div class="desktop-image">
-      {% image image, "", site.products.header_image_widths, "banner", "", "3.5/1", "eager" %}
+      {% image image, "", config.products.header_image_widths, "banner", "", "3.5/1", "eager" %}
     </div>
     <div class="mobile-image">
-      {% image image, "", site.products.header_image_widths, "banner", "", "1/1", "eager" %}
+      {% image image, "", config.products.header_image_widths, "banner", "", "1/1", "eager" %}
     </div>
   {% endif %}
 </header>
