+++ src/_includes/menu-list-item.html
@@ -1,7 +1,7 @@
 <li>
   {% assign item = menu %}
-  {% assign widths = site.products.category_thumb_widths %}
-  {% assign sizes = site.products.category_thumb_sizes %}
+  {% assign widths = config.products.category_thumb_widths %}
+  {% assign sizes = config.products.category_thumb_sizes %}
   {% include "thumbnail-image-link.html" %}
 
   {% if menu.data.description -%}
