+++ src/_includes/navigation.html
@@ -1,5 +1,5 @@
 <nav>
-  {%- if site.sticky_mobile_nav -%}
+  {%- if config.sticky_mobile_nav -%}
     <input
       type="checkbox"
       id="nav-toggle"
