+++ src/_layouts/home.html
@@ -4,14 +4,14 @@ layout: base.html
 
 {{ content }}
 
-{%- if site.homepage_news and collections.news.size > 0 -%}
+{%- if config.homepage_news and collections.news.size > 0 -%}
   <hr />
   {%- render_snippet "homepage-news-header", "<h3>Latest Posts</h3>" -%}
   {% include "news-post-list.html", limit: 2 %}
 {%- endif -%}
 
 {%- assign categories = collections.categories | getFeaturedCategories -%}
-{%- if site.homepage_products and categories.size > 0 -%}
+{%- if config.homepage_products and categories.size > 0 -%}
   <hr>
   {%- render_snippet "homepage-categories-header", "<h3>Product Categories</h3>" -%}
   {%- include "categories-list.html" -%}
