+++ src/_layouts/product-gallery.html
@@ -18,4 +18,4 @@ layout: base.html
   {{ page.imageName | escape }}
 </h2>
 
-{% image page.imageSrc, page.imageName, site.products.gallery_image_widths %}
+{% image page.imageSrc, page.imageName, config.products.gallery_image_widths %}
